+++ src/commands/mod.rs
@@ -1,3 +1,4 @@
 pub mod music;
 pub mod utility;
 pub mod help;
+pub mod reputation;
