+++ src/commands/utility/mod.rs
@@ -2,4 +2,3 @@ pub mod cmd_uwu;
 pub mod cmd_notify;
 pub mod cmd_wakeup;
 pub mod cmd_rename;
-pub mod cmd_rep;
